Pull out reusable parts of BeamFileSystemArtifact*.java into abstract superclass.
