+++ runners/java-fn-execution/src/main/java/org/apache/beam/runners/fnexecution/artifact/BeamFileSystemArtifactStagingService.java
@@ -17,38 +17,22 @@
  */
 package org.apache.beam.runners.fnexecution.artifact;
 
-import static org.apache.beam.vendor.guava.v26_0_jre.com.google.common.base.Preconditions.checkNotNull;
-
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.io.IOException;
 import java.io.Serializable;
 import java.nio.channels.WritableByteChannel;
-import java.nio.charset.Charset;
-import java.nio.charset.StandardCharsets;
 import java.util.Collections;
-import org.apache.beam.model.jobmanagement.v1.ArtifactApi.ArtifactMetadata;
-import org.apache.beam.model.jobmanagement.v1.ArtifactApi.CommitManifestRequest;
-import org.apache.beam.model.jobmanagement.v1.ArtifactApi.CommitManifestResponse;
 import org.apache.beam.model.jobmanagement.v1.ArtifactApi.ProxyManifest;
 import org.apache.beam.model.jobmanagement.v1.ArtifactApi.ProxyManifest.Location;
-import org.apache.beam.model.jobmanagement.v1.ArtifactApi.PutArtifactMetadata;
-import org.apache.beam.model.jobmanagement.v1.ArtifactApi.PutArtifactRequest;
-import org.apache.beam.model.jobmanagement.v1.ArtifactApi.PutArtifactResponse;
 import org.apache.beam.model.jobmanagement.v1.ArtifactStagingServiceGrpc.ArtifactStagingServiceImplBase;
-import org.apache.beam.runners.fnexecution.FnService;
 import org.apache.beam.sdk.io.FileSystems;
 import org.apache.beam.sdk.io.fs.MoveOptions.StandardMoveOptions;
 import org.apache.beam.sdk.io.fs.ResolveOptions.StandardResolveOptions;
 import org.apache.beam.sdk.io.fs.ResourceId;
 import org.apache.beam.sdk.util.MimeTypes;
-import org.apache.beam.vendor.grpc.v1p21p0.com.google.protobuf.ByteString;
-import org.apache.beam.vendor.grpc.v1p21p0.com.google.protobuf.util.JsonFormat;
 import org.apache.beam.vendor.grpc.v1p21p0.io.grpc.Status;
 import org.apache.beam.vendor.grpc.v1p21p0.io.grpc.StatusRuntimeException;
-import org.apache.beam.vendor.grpc.v1p21p0.io.grpc.stub.StreamObserver;
-import org.apache.beam.vendor.guava.v26_0_jre.com.google.common.hash.Hasher;
-import org.apache.beam.vendor.guava.v26_0_jre.com.google.common.hash.Hashing;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -66,86 +50,37 @@ import org.slf4j.LoggerFactory;
  *
  * <p>The manifest file is encoded in {@link ProxyManifest}.
  */
-public class BeamFileSystemArtifactStagingService extends ArtifactStagingServiceImplBase
-    implements FnService {
+public class BeamFileSystemArtifactStagingService extends AbstractArtifactStagingService {
 
   private static final Logger LOG =
       LoggerFactory.getLogger(BeamFileSystemArtifactStagingService.class);
   private static final ObjectMapper MAPPER = new ObjectMapper();
   // Use UTF8 for all text encoding.
-  private static final Charset CHARSET = StandardCharsets.UTF_8;
   public static final String MANIFEST = "MANIFEST";
   public static final String ARTIFACTS = "artifacts";
 
   @Override
-  public StreamObserver<PutArtifactRequest> putArtifact(
-      StreamObserver<PutArtifactResponse> responseObserver) {
-    return new PutArtifactStreamObserver(responseObserver);
+  public String getArtifactUri(String stagingSession, String encodedFileName) throws Exception {
+    StagingSessionToken stagingSessionToken = StagingSessionToken.decode(stagingSession);
+    ResourceId artifactDirResourceId = getArtifactDirResourceId(stagingSessionToken);
+    return artifactDirResourceId
+        .resolve(encodedFileName, StandardResolveOptions.RESOLVE_FILE)
+        .toString();
   }
 
   @Override
-  public void commitManifest(
-      CommitManifestRequest request, StreamObserver<CommitManifestResponse> responseObserver) {
-    try {
-      StagingSessionToken stagingSessionToken =
-          StagingSessionToken.decode(request.getStagingSessionToken());
-      ResourceId manifestResourceId = getManifestFileResourceId(stagingSessionToken);
-      ResourceId artifactDirResourceId = getArtifactDirResourceId(stagingSessionToken);
-      ProxyManifest.Builder proxyManifestBuilder =
-          ProxyManifest.newBuilder().setManifest(request.getManifest());
-      for (ArtifactMetadata artifactMetadata : request.getManifest().getArtifactList()) {
-        proxyManifestBuilder.addLocation(
-            Location.newBuilder()
-                .setName(artifactMetadata.getName())
-                .setUri(
-                    artifactDirResourceId
-                        .resolve(
-                            encodedFileName(artifactMetadata), StandardResolveOptions.RESOLVE_FILE)
-                        .toString())
-                .build());
-      }
-      try (WritableByteChannel manifestWritableByteChannel =
-          FileSystems.create(manifestResourceId, MimeTypes.TEXT)) {
-        manifestWritableByteChannel.write(
-            CHARSET.encode(JsonFormat.printer().print(proxyManifestBuilder.build())));
-      }
-      // TODO: Validate integrity of staged files.
-      responseObserver.onNext(
-          CommitManifestResponse.newBuilder()
-              .setRetrievalToken(manifestResourceId.toString())
-              .build());
-      responseObserver.onCompleted();
-    } catch (Exception e) {
-      // TODO: Cleanup all the artifacts.
-      LOG.error("Unable to commit manifest.", e);
-      responseObserver.onError(e);
-    }
+  public WritableByteChannel openUri(String uri) throws IOException {
+    return FileSystems.create(FileSystems.matchNewResource(uri, false), MimeTypes.BINARY);
   }
 
   @Override
-  public void close() throws Exception {
-    // Nothing to close here.
+  public void removeUri(String uri) throws IOException {
+    FileSystems.delete(
+        Collections.singletonList(FileSystems.matchNewResource(uri, false)),
+        StandardMoveOptions.IGNORE_MISSING_FILES);
   }
 
-  /**
-   * Generate a stagingSessionToken compatible with {@link BeamFileSystemArtifactStagingService}.
-   *
-   * @param sessionId Unique sessionId for artifact staging.
-   * @param basePath Base path to upload artifacts.
-   * @return Encoded stagingSessionToken.
-   */
-  public static String generateStagingSessionToken(String sessionId, String basePath) {
-    StagingSessionToken stagingSessionToken = new StagingSessionToken();
-    stagingSessionToken.setSessionId(sessionId);
-    stagingSessionToken.setBasePath(basePath);
-    return stagingSessionToken.encode();
-  }
-
-  private String encodedFileName(ArtifactMetadata artifactMetadata) {
-    return "artifact_"
-        + Hashing.sha256().hashString(artifactMetadata.getName(), CHARSET).toString();
-  }
-
+  @Override
   public void removeArtifacts(String stagingSessionToken) throws Exception {
     StagingSessionToken parsedToken = StagingSessionToken.decode(stagingSessionToken);
     ResourceId dir = getJobDirResourceId(parsedToken);
@@ -176,6 +111,19 @@ public class BeamFileSystemArtifactStagingService extends ArtifactStagingService
     LOG.info("Removed dir {}", dir);
   }
 
+  @Override
+  public WritableByteChannel openManifest(String stagingSession) throws Exception {
+    return FileSystems.create(
+        getManifestFileResourceId(StagingSessionToken.decode(stagingSession)), MimeTypes.TEXT);
+  }
+
+  @Override
+  public String getRetrievalToken(String stagingSession) throws Exception {
+    StagingSessionToken stagingSessionToken = StagingSessionToken.decode(stagingSession);
+    ResourceId manifestResourceId = getManifestFileResourceId(stagingSessionToken);
+    return manifestResourceId.toString();
+  }
+
   private ResourceId getJobDirResourceId(StagingSessionToken stagingSessionToken) {
     ResourceId baseResourceId;
     // Get or Create the base path
@@ -196,126 +144,25 @@ public class BeamFileSystemArtifactStagingService extends ArtifactStagingService
         .resolve(ARTIFACTS, StandardResolveOptions.RESOLVE_DIRECTORY);
   }
 
-  private class PutArtifactStreamObserver implements StreamObserver<PutArtifactRequest> {
-
-    private final StreamObserver<PutArtifactResponse> outboundObserver;
-    private PutArtifactMetadata metadata;
-    private ResourceId artifactId;
-    private WritableByteChannel artifactWritableByteChannel;
-    private Hasher hasher;
-
-    PutArtifactStreamObserver(StreamObserver<PutArtifactResponse> outboundObserver) {
-      this.outboundObserver = outboundObserver;
-    }
-
-    @Override
-    public void onNext(PutArtifactRequest putArtifactRequest) {
-      // Create the directory structure for storing artifacts in the first call.
-      if (metadata == null) {
-        checkNotNull(putArtifactRequest);
-        checkNotNull(putArtifactRequest.getMetadata());
-        metadata = putArtifactRequest.getMetadata();
-        LOG.debug("stored metadata: {}", metadata);
-        // Check the base path exists or create the base path
-        try {
-          ResourceId artifactsDirId =
-              getArtifactDirResourceId(
-                  StagingSessionToken.decode(
-                      putArtifactRequest.getMetadata().getStagingSessionToken()));
-          artifactId =
-              artifactsDirId.resolve(
-                  encodedFileName(metadata.getMetadata()), StandardResolveOptions.RESOLVE_FILE);
-          LOG.debug(
-              "Going to stage artifact {} to {}.", metadata.getMetadata().getName(), artifactId);
-          artifactWritableByteChannel = FileSystems.create(artifactId, MimeTypes.BINARY);
-          hasher = Hashing.sha256().newHasher();
-        } catch (Exception e) {
-          String message =
-              String.format(
-                  "Failed to begin staging artifact %s", metadata.getMetadata().getName());
-          LOG.error(message, e);
-          outboundObserver.onError(
-              new StatusRuntimeException(Status.DATA_LOSS.withDescription(message).withCause(e)));
-        }
-      } else {
-        try {
-          ByteString data = putArtifactRequest.getData().getData();
-          artifactWritableByteChannel.write(data.asReadOnlyByteBuffer());
-          hasher.putBytes(data.toByteArray());
-        } catch (IOException e) {
-          String message =
-              String.format(
-                  "Failed to write chunk of artifact %s to %s",
-                  metadata.getMetadata().getName(), artifactId);
-          LOG.error(message, e);
-          outboundObserver.onError(
-              new StatusRuntimeException(Status.DATA_LOSS.withDescription(message).withCause(e)));
-        }
-      }
-    }
-
-    @Override
-    public void onError(Throwable throwable) {
-      // Delete the artifact.
-      LOG.error("Staging artifact failed for " + artifactId, throwable);
-      try {
-        if (artifactWritableByteChannel != null) {
-          artifactWritableByteChannel.close();
-        }
-        if (artifactId != null) {
-          FileSystems.delete(
-              Collections.singletonList(artifactId), StandardMoveOptions.IGNORE_MISSING_FILES);
-        }
-
-      } catch (IOException e) {
-        outboundObserver.onError(
-            new StatusRuntimeException(
-                Status.DATA_LOSS.withDescription(
-                    String.format("Failed to clean up artifact file %s", artifactId))));
-        return;
-      }
-      outboundObserver.onError(
-          new StatusRuntimeException(
-              Status.DATA_LOSS
-                  .withDescription(String.format("Failed to stage artifact %s", artifactId))
-                  .withCause(throwable)));
-    }
-
-    @Override
-    public void onCompleted() {
-      // Close the stream.
-      LOG.debug("Staging artifact completed for " + artifactId);
-      if (artifactWritableByteChannel != null) {
-        try {
-          artifactWritableByteChannel.close();
-        } catch (IOException e) {
-          onError(e);
-          return;
-        }
-      }
-      String expectedSha256 = metadata.getMetadata().getSha256();
-      if (expectedSha256 != null && !expectedSha256.isEmpty()) {
-        String actualSha256 = hasher.hash().toString();
-        if (!actualSha256.equals(expectedSha256)) {
-          outboundObserver.onError(
-              new StatusRuntimeException(
-                  Status.INVALID_ARGUMENT.withDescription(
-                      String.format(
-                          "Artifact %s is corrupt: expected sah256 %s, but has sha256 %s",
-                          metadata.getMetadata().getName(), expectedSha256, actualSha256))));
-          return;
-        }
-      }
-      outboundObserver.onNext(PutArtifactResponse.newBuilder().build());
-      outboundObserver.onCompleted();
-    }
+  /**
+   * Generate a stagingSessionToken compatible with {@link BeamFileSystemArtifactStagingService}.
+   *
+   * @param sessionId Unique sessionId for artifact staging.
+   * @param basePath Base path to upload artifacts.
+   * @return Encoded stagingSessionToken.
+   */
+  public static String generateStagingSessionToken(String sessionId, String basePath) {
+    StagingSessionToken stagingSessionToken = new StagingSessionToken();
+    stagingSessionToken.setSessionId(sessionId);
+    stagingSessionToken.setBasePath(basePath);
+    return stagingSessionToken.encode();
   }
 
   /**
    * Serializable StagingSessionToken used to stage files with {@link
    * BeamFileSystemArtifactStagingService}.
    */
-  private static class StagingSessionToken implements Serializable {
+  protected static class StagingSessionToken implements Serializable {
 
     private String sessionId;
     private String basePath;
